+++ src/ray/common/test_util.cc
@@ -19,13 +19,13 @@
 
 #include "absl/strings/escaping.h"
 #include "ray/common/buffer.h"
-#include "ray/common/network_util.h"
 #include "ray/common/ray_config.h"
 #include "ray/common/ray_object.h"
 #include "ray/common/test_util.h"
 #include "ray/util/cmd_line_utils.h"
 #include "ray/util/filesystem.h"
 #include "ray/util/logging.h"
+#include "ray/util/network_util.h"
 #include "ray/util/path_utils.h"
 #include "ray/util/process.h"
 #include "ray/util/util.h"
