+++ src/ray/raylet/BUILD.bazel
@@ -108,7 +108,6 @@ ray_cc_library(
         ":runtime_env_agent_client",
         ":worker",
         "//src/ray/common:constants",
-        "//src/ray/common:network_util",
         "//src/ray/common:ray_config",
         "//src/ray/common:runtime_env",
         "//src/ray/common:status",
@@ -116,6 +115,7 @@ ray_cc_library(
         "//src/ray/core_worker:core_worker_common",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/ipc:client_connection",
+        "//src/ray/util:network_util",
         "@boost//:system",
         "@com_google_absl//absl/strings",
     ],
