+++ src/ray/raylet/worker_pool.cc
@@ -28,7 +28,6 @@
 
 #include "absl/strings/str_split.h"
 #include "ray/common/constants.h"
-#include "ray/common/network_util.h"
 #include "ray/common/ray_config.h"
 #include "ray/common/runtime_env_common.h"
 #include "ray/common/status.h"
@@ -37,6 +36,7 @@
 #include "ray/gcs/pb_util.h"
 #include "ray/stats/metric_defs.h"
 #include "ray/util/logging.h"
+#include "ray/util/network_util.h"
 #include "ray/util/util.h"
 
 DEFINE_stats(worker_register_time_ms,
